Support for auto-resuming timedout CI minimization
To have GH actions ask coqbot to auto-resume, the bot should stick the
resumption endpoint in `coqbot.resume-minimization-url`, stick the
docker image in `coqbot.docker-image` (in addition to
`.github/workflows/main.yml`), and should expect a reply with the
following lines:
1. `cat coqbot-request-stamp`
2. the commit message with newlines replaced by `\t` (useful for making a
   new informative commit message) (generated by `echo -n "$(git log -1
   --pretty=%B)" | tr '\n' '\t'`)
3. `cat coqbot.docker-image`
4. `cat coqbot.ci-target`
5. `cat coqbot.compiler`
6. `cat coqbot.failing-artifact-urls`
7. `cat coqbot.passing-artifact-urls`
8. `cat coqbot.failing-sha`
9. `cat coqbot.passing-sha`

After this (lines 10 and beyond) are the partially minimized file, which
should be placed in `bug.v` to resume minimization.

Note that we currently unconditionally comment with the partially
minimized file on GitHub, but we might want to change this in the
future?

c.f. https://github.com/coq/bot/issues/143#issuecomment-857700256